Refactor delegator balances
